+++ packages/grid/_modules_/grid/components/containers/GridToolbarContainer.tsx
@@ -1,30 +1,26 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { getDataGridUtilityClass } from '../../gridClasses';
-import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
-import { composeClasses } from '../../utils/material-ui-utils';
-import { GridComponentProps } from '../../GridComponentProps';
+import { makeStyles } from '@material-ui/styles';
+import { createTheme } from '../../utils/utils';
+
+const defaultTheme = createTheme();
+const useStyles = makeStyles(
+  {
+    root: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '4px 4px 0',
+    },
+  },
+  { name: 'MuiGridToolbarContainer', defaultTheme },
+);
 
 export type GridToolbarContainerProps = React.HTMLAttributes<HTMLDivElement>;
-
-type OwnerState = { classes: GridComponentProps['classes'] };
-
-const useUtilityClasses = (ownerState: OwnerState) => {
-  const { classes } = ownerState;
-
-  const slots = {
-    root: ['toolbarContainer'],
-  };
-
-  return composeClasses(slots, getDataGridUtilityClass, classes);
-};
 
 export const GridToolbarContainer = React.forwardRef<HTMLDivElement, GridToolbarContainerProps>(
   function GridToolbarContainer(props, ref) {
     const { className, children, ...other } = props;
-    const rootProps = useGridRootProps();
-    const ownerState = { classes: rootProps.classes };
-    const classes = useUtilityClasses(ownerState);
+    const classes = useStyles();
 
     if (!children) {
       return null;
