--- packages/grid/_modules_/grid/utils/nameof.ts
@@ -1,4 +1,0 @@
-/**
- * Ensures that a `name` is a valid identifier belonging to a `T` type.
- */
-export const nameof = <T>(name: keyof T) => name;
